[hist] Cast benchmark ranges
This avoids sign conversion warnings.

+++ hist/histv7/benchmark/hist_benchmark_axes.cxx
@@ -2,6 +2,7 @@
 
 #include <benchmark/benchmark.h>
 
+#include <cstddef>
 #include <random>
 #include <tuple>
 #include <vector>
@@ -19,7 +20,7 @@ struct RAxes_Regular1 : public benchmark::Fixture {
    {
       std::mt19937 gen;
       std::uniform_real_distribution<> dis;
-      fNumbers.resize(state.range(0));
+      fNumbers.resize(static_cast<std::size_t>(state.range(0)));
       for (std::size_t i = 0; i < fNumbers.size(); i++) {
          fNumbers[i] = dis(gen);
       }
@@ -49,7 +50,7 @@ struct RAxes_Regular2 : public benchmark::Fixture {
    {
       std::mt19937 gen;
       std::uniform_real_distribution<> dis;
-      fNumbers.resize(2 * state.range(0));
+      fNumbers.resize(2 * static_cast<std::size_t>(state.range(0)));
       for (std::size_t i = 0; i < fNumbers.size(); i++) {
          fNumbers[i] = dis(gen);
       }
